+++ test/view/scene_view_interactive_test.dart
@@ -12,6 +12,7 @@ import 'package:iwb_canvas_engine/src/interactive/scene_controller.dart';
 import 'package:iwb_canvas_engine/src/interactive/internal/pointer_session_token.dart';
 import 'package:iwb_canvas_engine/src/interactive/internal/scene_controller_pointer_session.dart';
 import 'package:iwb_canvas_engine/src/core/scene_snapshot_paint_candidates.dart';
+import 'package:iwb_canvas_engine/src/core/snapshot_paint_admission_bounds.dart';
 import 'package:iwb_canvas_engine/src/render/scene_painter.dart';
 import 'package:iwb_canvas_engine/src/view/scene_view_interactive_overlay_painter.dart';
 import 'package:iwb_canvas_engine/src/view/scene_view_interactive.dart';
@@ -2583,6 +2584,8 @@ class _StaticSceneViewReadState extends ChangeNotifier
   _StaticSceneViewReadState(this._snapshot);
 
   final SceneSnapshot _snapshot;
+  final SnapshotPaintAdmissionBoundsCache _snapshotPaintAdmissionBoundsCache =
+      SnapshotPaintAdmissionBoundsCache();
 
   @override
   SceneSnapshot get snapshot => _snapshot;
@@ -2623,6 +2626,7 @@ class _StaticSceneViewReadState extends ChangeNotifier
         query: query,
         selectedNodeIds: frameRead.selectedNodeIds,
         preview: frameRead.preview,
+        admissionBounds: _snapshotPaintAdmissionBoundsCache,
       ),
     );
   }